refactor: port 설정 및 securityConfig api 경로 설정 (#10)
