+++ auth/src/test/java/com/food/auth/integration/SecurityApiTest.java
@@ -1,0 +1,30 @@
+package com.food.auth.integration;
+
+import com.food.AuthApplication;
+import org.junit.jupiter.api.BeforeEach;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.web.servlet.MockMvc;
+import org.springframework.test.web.servlet.setup.MockMvcBuilders;
+import org.springframework.transaction.annotation.Transactional;
+import org.springframework.web.context.WebApplicationContext;
+
+import static org.springframework.security.test.web.servlet.setup.SecurityMockMvcConfigurers.springSecurity;
+
+@SpringBootTest(classes = AuthApplication.class)
+@Transactional
+public class SecurityApiTest {
+
+    @Autowired
+    private WebApplicationContext context;
+
+    protected MockMvc mvc;
+
+    @BeforeEach
+    void setup() {
+        mvc = MockMvcBuilders
+                .webAppContextSetup(context)
+                .apply(springSecurity())
+                .build();
+    }
+}
