+++ user/src/main/java/com/food/UserApplication.java
@@ -1,4 +1,4 @@
-package com.food.user;
+package com.food;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
